+++ test/up-cli.test.js
@@ -13,10 +13,14 @@
 /* eslint-env mocha */
 import sinon from 'sinon';
 import dotenv from 'dotenv';
+import fs from 'fs/promises';
+import { UnsecuredJWT } from 'jose';
+
 import path from 'path';
 import { clearHelixEnv } from './utils.js';
 import CLI from '../src/cli.js';
 import UpCommand from '../src/up.cmd.js';
+import { saveSiteTokenToFile } from '../src/config/config-utils.js';
 
 describe('hlx up', () => {
   // mocked command instance
@@ -45,6 +49,7 @@ describe('hlx up', () => {
     mockUp.withAllowInsecure.returnsThis();
     mockUp.withKill.returnsThis();
     mockUp.withCache.returnsThis();
+    mockUp.withSiteToken.returnsThis();
     mockUp.run.returnsThis();
     cli = (await new CLI().initCommands()).withCommandExecutor('up', mockUp);
   });
@@ -72,7 +77,19 @@ describe('hlx up', () => {
     sinon.assert.calledWith(mockUp.withHttpPort, 1234);
     sinon.assert.calledWith(mockUp.withBindAddr, '*');
     sinon.assert.calledWith(mockUp.withPrintIndex, true);
+    sinon.assert.calledWith(mockUp.withSiteToken, 'secret-site-token');
     sinon.assert.calledOnce(mockUp.run);
+  });
+
+  it('hlx up reads token from .hlx-token', async () => {
+    try {
+      const mockToken = `hlxtst_${new UnsecuredJWT({ email: 'test@example.com' }).encode()}`;
+      await saveSiteTokenToFile(mockToken);
+      await cli.run(['up']);
+      sinon.assert.calledWith(mockUp.withSiteToken, mockToken);
+    } finally {
+      await fs.rm(path.resolve(__rootdir, '.hlx'), { recursive: true, force: true });
+    }
   });
 
   it('hlx up fails with non env extra argument', async () => {
@@ -137,6 +154,12 @@ describe('hlx up', () => {
   it('hlx up can disable kill', async () => {
     await cli.run(['up', '--stop-other', 'false']);
     sinon.assert.calledWith(mockUp.withKill, false);
+    sinon.assert.calledOnce(mockUp.run);
+  });
+
+  it('aem up can set site token', async () => {
+    await cli.run(['up', '--site-token', 'secret-site-token']);
+    sinon.assert.calledWith(mockUp.withSiteToken, 'secret-site-token');
     sinon.assert.calledOnce(mockUp.run);
   });
 });